pyproject.toml
[tool.poetry]
name = "create-react-django-app"
version = "0.1.0"
description = "Minimal Django + Create React App template"
authors = ["kevinshen56714 <kshen64@gatech.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
psycopg2 = "^2.9.3"
python-decouple = "^3.6"
django-filter = "^23.1"
celery = "^5.2.7"
django-celery-beat = "^2.5.0"
gevent = "^22.10.2"
sqlalchemy = "^2.0.9"
requests = "^2.28.2"
rpi-gpio = "^0.7.1"
redis = "^4.5.4"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
